smart-writer/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1
html2text>=2020.1.16
google-generativeai>=0.8.0
python-dotenv>=1.0.0
